# backend/internal/httpapi/panelsettings/docs/swagger.yaml
basePath: /api
definitions:
...
  shared.ErrorResponse:
    properties:
      error:
        example: invalid input
        type: string
      message:
        example: invalid input
        type: string
    type: object
...
info:
  contact: {}
  description: Exodus dashboard and node-management API.
  license:
    name: AGPL-3.0
  title: Exodus API
  version: 3.0.0
paths:
...
  /srs-lists:
    get:
      consumes:
      - application/json
      description: List, create (201), or update sing-box binary rule-set lists
      parameters:
      - description: SRS rule list payload
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Manage SRS rule lists
      tags:
      - SRS Lists Controller
    patch:
      consumes:
      - application/json
      description: List, create (201), or update sing-box binary rule-set lists
      parameters:
      - description: SRS rule list payload
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Manage SRS rule lists
      tags:
      - SRS Lists Controller
    post:
      consumes:
      - application/json
      description: List, create (201), or update sing-box binary rule-set lists
      parameters:
      - description: SRS rule list payload
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Manage SRS rule lists
      tags:
      - SRS Lists Controller
  /srs-lists/{uuid}:
    delete:
      description: Get or delete SRS rule list by UUID
      parameters:
      - description: SRS rule list UUID
        format: uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: SRS rule list by UUID
      tags:
      - SRS Lists Controller
    get:
      description: Get or delete SRS rule list by UUID
      parameters:
      - description: SRS rule list UUID
        format: uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: SRS rule list by UUID
      tags:
      - SRS Lists Controller
  /srs-lists/actions/check:
    post:
      consumes:
      - application/json
      description: Reorder, check or trigger subscription node sync for SRS rule lists
      parameters:
      - description: Action payload
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: SRS rule list actions
      tags:
      - SRS Lists Controller
  /srs-lists/actions/reorder:
    post:
      consumes:
      - application/json
      description: Reorder, check or trigger subscription node sync for SRS rule lists
      parameters:
      - description: Action payload
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: SRS rule list actions
      tags:
      - SRS Lists Controller
  /srs-lists/actions/sync:
    post:
      consumes:
      - application/json
      description: Reorder, check or trigger subscription node sync for SRS rule lists
      parameters:
      - description: Action payload
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: SRS rule list actions
      tags:
      - SRS Lists Controller
  /srs-lists/bulk/delete:
    post:
      consumes:
      - application/json
      description: Bulk delete, enable, disable, or update download interval for SRS
        lists
      parameters:
      - description: Bulk payload
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Bulk SRS rule list operations
      tags:
      - SRS Lists Controller
  /srs-lists/bulk/disable:
    post:
      consumes:
      - application/json
      description: Bulk delete, enable, disable, or update download interval for SRS
        lists
      parameters:
      - description: Bulk payload
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Bulk SRS rule list operations
      tags:
      - SRS Lists Controller
  /srs-lists/bulk/enable:
    post:
      consumes:
      - application/json
      description: Bulk delete, enable, disable, or update download interval for SRS
        lists
      parameters:
      - description: Bulk payload
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Bulk SRS rule list operations
      tags:
      - SRS Lists Controller
  /srs-lists/bulk/set-interval:
    post:
      consumes:
      - application/json
      description: Bulk delete, enable, disable, or update download interval for SRS
        lists
      parameters:
      - description: Bulk payload
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Bulk SRS rule list operations
      tags:
      - SRS Lists Controller
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"